[REFACTOR] move demo code in a dedicated folder
This explicitly shows that it is not part of the lib.
